Widget class and display options moved into their own files
